Bump physis pin to upstream main plus our excel fixes
Point the workspace patch at Yarukon/Physis db33aec (origin/main plus
subrow offset and SeString decoding). TreasureInstanceObject now nests
base_id under parent_data as u32, so the zone treasure cache and its
lookup follow that layout.

+++ servers/world/src/server/director.rs
@@ -2221,7 +2221,7 @@ pub fn director_tick(
                 }
             }
             LuaDirectorTask::SpawnTreasure { id } => {
-                if let Some(mut treasure) = instance.zone.get_treasure(*id as u8) {
+                if let Some(mut treasure) = instance.zone.get_treasure(*id) {
                     treasure.handler_id = director_id;
 
                     let actor_id = ObjectId(fastrand::u32(..));
